+++ tests/testthat/resources/knitr-spin.R
@@ -1,0 +1,8 @@
+#' ---
+#' title: Palmer Penguins
+#' author: Norah Jones
+#' date: 3/12/23
+#' format: html
+#' ---
+
+print(42)